refactor(botonic-react): rename functions and file
